Make LiskAPI a class
