+++ src/api/liskApi.js
@@ -42,7 +42,6 @@
 import privateApi from './privateApi';
 import { checkOptions, optionallyCallCallback, wrapSendRequest } from './utils';
 import config from '../../config.json';
-import cryptoModule from '../crypto';
 
 const GET = 'GET';
 const POST = 'POST';
@@ -51,175 +50,174 @@ const livePort = 8000;
 const testPort = 7000;
 const sslPort = 443;
 
-function LiskAPI(providedOptions = {}) {
-	if (!(this instanceof LiskAPI)) {
-		return new LiskAPI(providedOptions);
-	}
-
-	const options = Object.assign({}, config.options, providedOptions);
-	const getDefaultPort = () => {
-		if (options.testnet) return testPort;
-		if (options.ssl) return sslPort;
-		return livePort;
-	};
-
-	this.defaultPeers = options.peers || config.peers.mainnet;
-
-	this.defaultSSLPeers = this.defaultPeers;
-
-	this.defaultTestnetPeers = options.peers || config.peers.testnet;
-
-	this.options = options;
-	this.ssl = options.ssl;
-	// Random peer can be set by settings with randomPeer: true | false
-	// Random peer is automatically enabled when no options.node has been entered. Else will be set
-	// to false.
-	// If the desired behaviour is to have an own node and automatic peer discovery, randomPeer
-	// should be set to true explicitly
-	this.randomPeer = (typeof options.randomPeer === 'boolean') ? options.randomPeer : !(options.node);
-	this.testnet = options.testnet;
-	this.bannedPeers = options.bannedPeers;
-	this.currentPeer = options.node || privateApi.selectNode.call(this);
-	this.port = (options.port === '' || options.port)
-		? options.port
-		: getDefaultPort(options);
-	this.nethash = this.getNethash(options.nethash);
+function getDefaultPort(options) {
+	if (options.testnet) return testPort;
+	if (options.ssl) return sslPort;
+	return livePort;
 }
 
-/**
- * @method getNethash
- * @return {object}
- * @public
- */
+class LiskAPI {
+	constructor(providedOptions) {
+		const options = Object.assign({}, config.options, providedOptions);
 
-LiskAPI.prototype.getNethash = function getNethash(providedNethash) {
-	const NetHash = this.testnet
-		? privateApi.netHashOptions.call(this).testnet
-		: privateApi.netHashOptions.call(this).mainnet;
+		this.defaultPeers = options.peers || config.peers.mainnet;
 
-	if (providedNethash) {
-		NetHash.nethash = providedNethash;
-		NetHash.version = '0.0.0a';
+		this.defaultSSLPeers = this.defaultPeers;
+
+		this.defaultTestnetPeers = options.peers || config.peers.testnet;
+
+		this.options = options;
+		this.ssl = options.ssl;
+		// Random peer can be set by settings with randomPeer: true | false
+		// Random peer is automatically enabled when no options.node has been entered. Else will be set
+		// to false.
+		// If the desired behaviour is to have an own node and automatic peer discovery, randomPeer
+		// should be set to true explicitly
+		this.randomPeer = (typeof options.randomPeer === 'boolean') ? options.randomPeer : !(options.node);
+		this.testnet = options.testnet;
+		this.bannedPeers = options.bannedPeers;
+		this.currentPeer = options.node || privateApi.selectNode.call(this);
+		this.port = (options.port === '' || options.port)
+			? options.port
+			: getDefaultPort(options);
+		this.nethash = this.getNethash(options.nethash);
 	}
 
-	return NetHash;
-};
+	/**
+	 * @method getNethash
+	 * @return {object}
+	 * @public
+	 */
 
-/**
- * @method getPeers
- * @return {object}
- */
+	getNethash(providedNethash) {
+		const NetHash = this.testnet
+			? privateApi.netHashOptions.call(this).testnet
+			: privateApi.netHashOptions.call(this).mainnet;
 
-LiskAPI.prototype.getPeers = function getPeers() {
-	return {
-		official: this.defaultPeers.map(node => ({ node })),
-		ssl: this.defaultSSLPeers.map(node => ({ node, ssl: true })),
-		testnet: this.defaultTestnetPeers.map(node => ({ node, testnet: true })),
-	};
-};
+		if (providedNethash) {
+			NetHash.nethash = providedNethash;
+			NetHash.version = '0.0.0a';
+		}
 
-/**
- * @method setNode
- * @param node string
- * @return {object}
- */
-
-LiskAPI.prototype.setNode = function setNode(node) {
-	this.currentPeer = node || privateApi.selectNode.call(this);
-	return this.currentPeer;
-};
-
-/**
- * @method setTestnet
- * @param testnet boolean
- */
-
-LiskAPI.prototype.setTestnet = function setTestnet(testnet) {
-	if (this.testnet !== testnet) {
-		this.bannedPeers = [];
+		return NetHash;
 	}
-	this.testnet = testnet;
-	this.port = testnet ? testPort : livePort;
 
-	privateApi.selectNode.call(this);
-};
+	/**
+	 * @method getPeers
+	 * @return {object}
+	 */
 
-/**
- * @method setSSL
- * @param ssl boolean
- */
+	getPeers() {
+		return {
+			official: this.defaultPeers.map(node => ({ node })),
+			ssl: this.defaultSSLPeers.map(node => ({ node, ssl: true })),
+			testnet: this.defaultTestnetPeers.map(node => ({ node, testnet: true })),
+		};
+	}
 
-LiskAPI.prototype.setSSL = function setSSL(ssl) {
-	if (this.ssl !== ssl) {
-		this.ssl = ssl;
-		this.bannedPeers = [];
+	/**
+	 * @method setNode
+	 * @param node string
+	 * @return {object}
+	 */
+
+	setNode(node) {
+		this.currentPeer = node || privateApi.selectNode.call(this);
+		return this.currentPeer;
+	}
+
+	/**
+	 * @method setTestnet
+	 * @param testnet boolean
+	 */
+
+	setTestnet(testnet) {
+		if (this.testnet !== testnet) {
+			this.bannedPeers = [];
+		}
+		this.testnet = testnet;
+		this.port = testnet ? testPort : livePort;
+
 		privateApi.selectNode.call(this);
 	}
-};
 
-/**
- * @method getAddressFromSecret
- * @param secret
- *
- * @return keys object
- */
+	/**
+	 * @method setSSL
+	 * @param ssl boolean
+	 */
 
-LiskAPI.prototype.getAddressFromSecret = function getAddressFromSecret(secret) {
-	const accountKeys = cryptoModule.getKeys(secret);
-	const accountAddress = cryptoModule.getAddress(accountKeys.publicKey);
+	setSSL(ssl) {
+		if (this.ssl !== ssl) {
+			this.ssl = ssl;
+			this.bannedPeers = [];
+			privateApi.selectNode.call(this);
+		}
+	}
 
-	return {
-		address: accountAddress,
-		publicKey: accountKeys.publicKey,
-	};
-};
+	/**
+	 * @method broadcastSignedTransaction
+	 * @param transaction
+	 * @param callback
+	 *
+	 * @return API object
+	 */
 
-/**
- * @method broadcastSignedTransaction
- * @param transaction
- * @param callback
- *
- * @return API object
- */
+	broadcastSignedTransaction(
+		transaction, callback,
+	) {
+		const request = {
+			requestUrl: `${privateApi.getFullURL.call(this)}/api/transactions`,
+			nethash: this.nethash,
+			requestParams: { transaction },
+		};
 
-LiskAPI.prototype.broadcastSignedTransaction = function broadcastSignedTransaction(
-	transaction, callback,
-) {
-	const request = {
-		requestUrl: `${privateApi.getFullURL.call(this)}/api/transactions`,
-		nethash: this.nethash,
-		requestParams: { transaction },
-	};
+		privateApi.sendRequestPromise.call(this, POST, request).then(result => callback(result.body));
+	}
 
-	privateApi.sendRequestPromise.call(this, POST, request).then(result => callback(result.body));
-};
+	/**
+	 * @method sendRequest
+	 * @param requestMethod
+	 * @param requestType
+	 * @param optionsOrCallback
+	 * @param callbackIfOptions
+	 *
+	 * @return APIanswer Object
+	 */
 
-/**
- * @method sendRequest
- * @param requestMethod
- * @param requestType
- * @param optionsOrCallback
- * @param callbackIfOptions
- *
- * @return APIanswer Object
- */
+	sendRequest(
+		requestMethod, requestType, optionsOrCallback, callbackIfOptions,
+	) {
+		const callback = callbackIfOptions || optionsOrCallback;
+		const options = (typeof optionsOrCallback !== 'function' && typeof optionsOrCallback !== 'undefined')
+			? checkOptions(optionsOrCallback)
+			: {};
 
-LiskAPI.prototype.sendRequest = function sendRequest(
-	requestMethod, requestType, optionsOrCallback, callbackIfOptions,
-) {
-	const callback = callbackIfOptions || optionsOrCallback;
-	const options = (typeof optionsOrCallback !== 'function' && typeof optionsOrCallback !== 'undefined')
-		? checkOptions(optionsOrCallback)
-		: {};
+		return privateApi.sendRequestPromise.call(this, requestMethod, requestType, options)
+			.then(result => result.body)
+			.then(privateApi.handleTimestampIsInFutureFailures.bind(
+				this, requestMethod, requestType, options,
+			))
+			.catch(privateApi.handleSendRequestFailures.bind(this, requestMethod, requestType, options))
+			.then(optionallyCallCallback.bind(null, callback));
+	}
 
-	return privateApi.sendRequestPromise.call(this, requestMethod, requestType, options)
-		.then(result => result.body)
-		.then(privateApi.handleTimestampIsInFutureFailures.bind(
-			this, requestMethod, requestType, options,
-		))
-		.catch(privateApi.handleSendRequestFailures.bind(this, requestMethod, requestType, options))
-		.then(optionallyCallCallback.bind(null, callback));
-};
+	/**
+	 * @method sendLSK
+	 * @param recipientId
+	 * @param amount
+	 * @param secret
+	 * @param secondSecret
+	 * @param callback
+	 *
+	 * @return API object
+	 */
+
+	sendLSK(
+		recipientId, amount, secret, secondSecret, callback,
+	) {
+		return this.sendRequest(POST, 'transactions', { recipientId, amount, secret, secondSecret }, callback);
+	}
+}
 
 /**
  * @method getAccount
@@ -386,21 +384,5 @@ LiskAPI.prototype.getDapps = wrapSendRequest(GET, 'dapps', data => data);
 
 LiskAPI.prototype.getDappsByCategory = wrapSendRequest(GET, 'dapps', category => ({ category }));
 
-/**
- * @method sendLSK
- * @param recipientId
- * @param amount
- * @param secret
- * @param secondSecret
- * @param callback
- *
- * @return API object
- */
-
-LiskAPI.prototype.sendLSK = function sendLSK(
-	recipientId, amount, secret, secondSecret, callback,
-) {
-	return this.sendRequest(POST, 'transactions', { recipientId, amount, secret, secondSecret }, callback);
-};
 
 module.exports = LiskAPI;
