+++ test/crypto/keys.js
@@ -16,6 +16,7 @@ import {
 	getPrivateAndPublicKeyFromSecret,
 	getRawPrivateAndPublicKeyFromSecret,
 	getKeys,
+	getAddressFromSecret,
 } from '../../src/crypto/keys';
 import {
 	bufferToHex,
@@ -23,6 +24,10 @@ import {
 
 describe('keys', () => {
 	const defaultSecret = 'secret';
+	const defaultAddress = {
+		publicKey: '5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09',
+		address: '18160565574430594874L',
+	};
 	const expectedPublicKey = '5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
 	const expectedPrivateKey = '2bb80d537b1da3e38bd30361aa855686bde0eacd7162fef6a25fe97bf527a25b5d036a858ce89f844491762eb89e2bfbd50a4a0a0da658e4b2628b25b117ae09';
 	describe('#getPrivateAndPublicKeyFromSecret', () => {
@@ -62,6 +67,13 @@ describe('keys', () => {
 				publicKey: expectedPublicKey,
 				privateKey: expectedPrivateKey,
 			});
+		});
+	});
+
+	describe('#getAddressFromSecret', () => {
+		it('should create correct address and publicKey', () => {
+			// (LSK.getAddressFromSecret(defaultSecret)).should.eql(defaultAddress);
+			(getAddressFromSecret(defaultSecret)).should.eql(defaultAddress);
 		});
 	});
 });
